Backend/requirements.txt
# Backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
dj-database-url==3.0.1
Django==5.2.7
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-extensions==4.1
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-extra-fields==3.7.0
filetype==1.2.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
kombu==5.5.4
lxml==6.0.2
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg-binary==3.2.11
psycopg2==2.9.11
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
